pyproject.toml
[tool.poetry]
name = "ya_tracker_client"
version = "0.1.0"
description = "Async Yandex Tracker Client"
authors = ["Дмитрий Анфимов <danfimov@yandex-team.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ya_tracker_client"}]

[[tool.poetry.source]]
name = "default-pypi"
url = "https://pypi.org/simple"
priority = "default"

[[tool.poetry.source]]
name = "yandex-pypi"
url = "https://pypi.yandex-team.ru/simple"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.5"
pydantic = "^2.3.0"
certifi = "^2023.7.22"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
coverage = "^7.3.1"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
ruff = "^0.0.287"
polyfactory = "^2.8.2"
gevent = "^23.9.1"
pre-commit = "^3.4.0"

[tool.poetry.group.examples.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
python_files = "test_*.py"
python_classes = ["*Test", "Test*"]
python_functions = "test_*"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "# pragma: no cover",
    "def __repr__",
    "def __str__",
    "def __unicode__",
]
show_missing = true
skip_empty = true
omit = [
  "*/__init__.py",
]

[tool.coverage.html]
directory = "pytest-cov-report"

[tool.coverage.run]
branch = true
concurrency = ['gevent']

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
select = ["W", "E", "F", "Q", "B", "I", "N", "ASYNC", "G", "RUF", "COM", "C90"]
ignore = []
target-version = "py311"

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.isort]
known-local-folder = ["ya_tacker_client"]
lines-after-imports = 2
